feat(batch-n): split EventHandler into per-handler files + Notifier module
- lib/EventHandler.ts: replaced 1233-line file with 11-line barrel re-export
- lib/handlers/: 11 new files, one per exported handler function;
  private helpers co-located with their handler
- lib/Notifier.ts: new module owning all 4 notification functions
  (notifyRocketChatUserAsync, notifyRocketChatUserInRoomAsync,
  generateHintMessageWithTeamsLoginButton, notifyNotLoggedInUserAsync)
- lib/MessageHelper.ts: removed 3 notification exports + trimmed
  unused imports (INotifier, IMessageAction, MessageActionType, LoginButtonText)
  515 -> 463 lines
- 8 caller files updated to import notifications from lib/Notifier
- tsc --noEmit passes clean
- docs/refactor-progress.md updated
